+++ src/aux-common/partitions/RemoteCausalRepoPartition.ts
@@ -59,6 +59,7 @@ import {
     StateUpdatedEvent,
     stateUpdatedEvent,
     BotsState,
+    ON_REMOTE_DATA_ACTION_NAME,
 } from '../bots';
 import { flatMap } from 'lodash';
 import {
@@ -526,6 +527,19 @@ export class RemoteCausalRepoPartitionImpl
                                     const remoteAction = event.action
                                         .event as ShoutAction;
                                     this._onEvents.next([
+                                        action(
+                                            ON_REMOTE_DATA_ACTION_NAME,
+                                            null,
+                                            null,
+                                            {
+                                                name: remoteAction.eventName,
+                                                that: remoteAction.argument,
+                                                playerId:
+                                                    event.action.device.claims[
+                                                        SESSION_ID_CLAIM
+                                                    ],
+                                            }
+                                        ),
                                         action(
                                             ON_REMOTE_WHISPER_ACTION_NAME,
                                             null,